Initial commit for refactoring skip model code using class factories and abstract classes.
